Fix stuck buttons when the frame loses focus.
If a game is paused by the frontend while a key is being pressed, the
key will remain active for the game when unpaused. This is an issue
because the key will likely be released outside of the game zone;
therefore, we would not process the key release for the game itself.

The same bug happens if clicking outside of the game zone while
holding a key.

For the case of a directional key, this means a continuous movement
in the direction of the key pressed before the pause until the key
is pressed and released.

- Fix #142.