Issue # 563 Build fix: When OpenAL is not installed, building with wxWidgets failed.
N.B. Not sure if it's a good fix. I don't know CMake.

The error message was:
---
CMake Error: The following variables are used in this project, but they are set to NOTFOUND.
Please set them or make sure they are set and tested correctly in the CMake files:
OPENAL_LIBRARY (ADVANCED)
---