Add support for overriding games options on code.
Since we need to set some options explicitly to play some hack ROMs,
we start a manual override function on code for some common hacks.

Our current file override has one big problem regarding indexes; it
uses the game initials as a single index. Most hacks/patches of a ROM
have the same key, and therefore it is not good enough for some cases.

Issue #484 was a case of the hack ROM wanting a save option, while the
original used another. If the user tried the GUI to override the option,
then the original game would start to fail to read/write the battery.

- Fix #484.